feat: enhance project configuration with modern setup
