--- .babelrc
@@ -1,10 +1,0 @@
-{
-    "presets": ["es2015", "stage-0"],
-    "env": {
-        "test": {
-            "plugins": [
-                "istanbul"
-            ]
-        }
-    }
-}
