+++ .eslintignore
@@ -1,0 +1,3 @@
+build
+dist
+node_modules
