--- .eslintrc
@@ -1,8 +1,0 @@
-{
-  "extends": "trendmicro",
-  "parser": "babel-eslint",
-  "env": {
-    "browser": true,
-    "node": true
-  }
-}
