+++ .gitignore
@@ -1,4 +1,4 @@
-node_modules
-/.nyc_output
-/lib
+/build
 /coverage
+/dist
+node_modules
