+++ .npmignore
@@ -1,1 +1,2 @@
-/coverage
+package-lock.json
+yarn.lock
