--- .travis.yml
@@ -1,7 +1,0 @@
-language: node_js
-node_js:
-  - '8'
-  - '6'
-  - '4'
-after_success:
-  - npm run coveralls
