+++ babel.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: '@trendmicro/babel-config',
+  presets: [
+    '@babel/preset-env',
+  ],
+};
