+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "gcode-toolpath",
-  "version": "2.2.0",
+  "version": "3.0.0",
   "description": "G-code Toolpath Generator",
-  "author": "Cheton Wu <cheton@gmail.com>",
+  "author": "Cheton Wu",
   "homepage": "https://github.com/cncjs/gcode-toolpath",
   "license": "MIT",
   "repository": {
@@ -17,43 +17,54 @@
     "gcode"
   ],
   "scripts": {
-    "prepublish": "npm run eslint && npm run build && npm test",
-    "eslint": "eslint src",
-    "build": "babel --out-dir ./lib ./src",
-    "test": "cross-env NODE_ENV=test nyc mocha",
-    "coveralls": "cat ./coverage/lcov.info | ./node_modules/.bin/coveralls"
+    "build": "cross-env rollup --config rollup.config.mjs",
+    "clean": "del build coverage dist",
+    "eslint": "eslint --ext .js,.jsx,.mjs .",
+    "pre-push": "bash -c 'echo -e \"=> \\e[1;33m$npm_package_name\\e[0m\"' && yarn run build && yarn run eslint && yarn run test",
+    "prepublish": "yarn run build",
+    "test": "jest --maxWorkers=2"
   },
-  "main": "lib/index.js",
+  "sideEffects": false,
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "files": [
+    "dist"
+  ],
   "dependencies": {
-    "gcode-interpreter": "^2.0.2"
+    "gcode-interpreter": "^3.0.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^8.0.3",
-    "babel-plugin-istanbul": "^4.1.5",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.26.0",
-    "chai": "^4.1.2",
-    "coveralls": "^3.0.0",
-    "cross-env": "^5.1.3",
-    "eslint": "^4.14.0",
-    "eslint-config-trendmicro": "^1.2.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.5.1",
-    "mocha": "^4.0.1",
-    "nyc": "^11.4.1"
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.0",
+    "@babel/eslint-parser": "^7.0.0",
+    "@babel/eslint-plugin": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@rollup/plugin-babel": "^6.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.0",
+    "@trendmicro/babel-config": "^1.0.2",
+    "cross-env": "^7.0.3",
+    "del-cli": "^5.0.0",
+    "eslint": "^8.25.0",
+    "jest": "^29.0.0",
+    "rollup": "^3"
   },
-  "nyc": {
-    "require": [
-      "babel-register"
+  "jest": {
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "<rootDir>/src/**/*.js"
     ],
-    "reporter": [
+    "coverageReporters": [
       "lcov",
-      "text"
+      "text",
+      "html"
     ],
-    "sourceMap": false,
-    "instrument": false
+    "modulePathIgnorePatterns": [],
+    "setupFiles": [],
+    "setupFilesAfterEnv": [],
+    "testEnvironment": "node",
+    "testMatch": [
+      "<rootDir>/**/__tests__/**/*.test.js"
+    ]
   }
 }
