+++ src/__tests__/fixtures/arc-xy-plane.nc
