+++ src/__tests__/fixtures/arc-yz-plane.nc
