+++ src/__tests__/fixtures/coordinate.nc
