+++ src/__tests__/fixtures/helical-thread-milling.nc
