+++ src/__tests__/fixtures/motion.nc
