+++ src/__tests__/fixtures/one-inch-circle.nc
